+++ inst/python/naiveBayesFunctions.py
@@ -81,7 +81,7 @@ def train_naive_bayes(population, plpData, modelOutput, variableNumber, quiet):
   if not os.path.exists(modelOutput):
     os.makedirs(modelOutput)
   print("Model saved to: %s" %(modelOutput)	)
-  joblib.dump(gnb, os.path.join(modelOutput,"model.pkl"))  
+  joblib.dump(gnb, os.path.join(modelOutput,"model.pkl"), compress = True)  
   # merge pred with indexes[testInd,:]
   test_pred = gnb.predict_proba(X)[:,1]
   test_pred.shape = (population[population[:,population.shape[1]-1] > 0,:].shape[0], 1)